requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# HuggingFace ecosystem
transformers>=4.35.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.24.0

# Logging and experiment tracking
wandb>=0.16.0
tensorboard>=2.14.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Evaluation metrics
nltk>=3.8.0
pycocoevalcap>=1.2

# Utilities
tqdm>=4.66.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
ipython>=8.12.0
jupyter>=1.0.0
